+++ RELEASE-NOTES.md
@@ -1,5 +1,9 @@
 
-# 1.8.34 (PLANNED)
+# 1.8.35 (PLANNED)
+
+TBD
+
+# 1.8.34
 
 * Add message source into ``ErrorV2``.
 * Deprecate ``TimeSource`` in brikar-common in favor of similar abstraction from ``utc-time`` library.